refactor welcome screen into its own element
